requirements.txt
# Data Processing & Computation
pandas>=2.0.0
numpy>=1.24.0
python-binance>=1.0.35

#config
PyYAML>=6.0.0

# Machine Learning & Evaluation
scikit-learn>=1.2.0
xgboost>=3.2.0

# Data Visualization
matplotlib>=3.7.0
